# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fitness Tracking Platform API
  version: 1.0.0
  description: Authoritative REST API for multi-user fitness management, daily tracking, workout/diet plan builders, progress analysis, reminders, and administration.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
paths:
...
  /me/today:
    get:
      summary: Resolve daily agenda and compliance tasks for today
      tags: [Daily Plan]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        '200':
          description: Resolved daily tasks and protocol targets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyPlanResponse'
  /me/days/{date}:
    get:
      summary: Resolve daily agenda for a specific historical or target date
      tags: [Daily Plan]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Resolved agenda for date
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DailyPlanResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            date: { type: string, format: date }
            timezone: { type: string }
            tasks:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  taskType: { type: string }
                  taskKey: { type: string }
                  title: { type: string }
                  status: { type: string }
                  isCompleted: { type: boolean }
            weight: { type: object }
            water: { type: object }
            diet: { type: object }
            workout: { type: object }
            cardio: { type: object }
            summary:
              type: object
              properties:
                completedTasks: { type: integer }
                totalTasks: { type: integer }
                overallAdherencePct: { type: number }